added some notes regarding printer installation
I've tried out installation of the Canon SELPHY CP 1500 using our old
Kano PI which is a Raspi 3 Model B Revision 1.2. After installing all
the needed packages (cups, gutenprint), the printer could simply be
selected in the CUPS configuration on localhost:631.
Borderless printing seems to be a little tricky, probably need to set
up a larger custom paper format which includes the margins.
Second "issue" is that it takes some time for the printing job to be
processed which adds to the ~30 seconds to finalize a picture.